fix: added HOC in index.tsx to handle onCloseEnd issues
